backend/requirements.txt
Flask==3.0.0
Flask_Cors==4.0.0
flask_sqlalchemy==3.1.1
flask_bcrypt==1.0.1
mysql-connector-python==8.2.0
PyMySQL==1.1.0
python-dotenv==1.0.0
flask-session
redis
itsdangerous==2.1.2
Flask-Mail==0.9.1